plugins/yihai_scrapy/requirements.txt
pyexecjs = 1.5.1
scrapy = 2.11.2
scrapy-redis = 0.9.1
psycopg2 = 2.9.9
pymysql = 1.1.1